requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.28.2
python-dotenv==1.0.0
PyPDF2==3.0.1
argparse==1.4.0